refactor: update CommitDetails component
